refactor(sign-up-age): PR changes
