+++ app/redux/types/cacheTypes.ts
@@ -6,14 +6,17 @@ export interface ISignUpData {
   avatar?: null;
 }
 
-export type IShortSignUpData = Partial<
-  Pick<ISignUpData, 'avatar' | 'nickname'> | { age?: string }
->;
+export interface IShortSignUpData {
+  nickname?: string;
+  age?: string;
+  // TODO: types
+  avatar?: null;
+}
 
 export interface ISignUpDataPayload {
-  payload: Partial<ISignUpData>;
+  payload: ISignUpData;
 }
 
 export interface IShortSignUpDataPayload {
-  payload: Partial<IShortSignUpData>;
+  payload: IShortSignUpData;
 }
