+++ app/screens/Onboarding/SignUp/SignUpAge/SignUpAge.styles.ts
